Add support for indefinite retention intervals
This adds support for a `retention_interval = :forever` option, which
will not expire any tables.

It also changes the migration to minimize downtime, by first copying
data to the `_new` table, and then swapping them at the end. It also
adds a `final_copy` rake task to facilitate copying any rows that were
inserted into the original table while the migration was happening.